requirements.txt
rich
pillow
pytest
